development-test/requirements-conda.txt
gdal==3.1.4
click==7.1.2
pyyaml==5.4.1
rasterio==1.2.0
setuptools==52.0.0
glob2==0.7
geopandas==0.8.2
matplotlib==3.3.4
h5py==3.1.0
appdirs==1.4.4
future==0.18.2
pystac==0.5.4
sentinelsat==0.14
python-dotenv==0.15.0
requests==2.25.1
urllib3==1.26.3
python-dotenv==0.15.0
sqlite==3.34.0